+++ internal/repository/user.go
@@ -5,44 +5,114 @@ import (
 	"errors"
 	"fmt"
 	"rest1/internal/domain"
+	"rest1/internal/usecases"
 	"sync"
 
 	"github.com/google/uuid"
 	// "github.com/jackc/pgx/v5"
+	"github.com/jackc/pgx/v5"
 	"github.com/jackc/pgx/v5/pgxpool"
 	"go.uber.org/zap"
 )
 
-type UserRepo struct {
+var (
+	// ErrTransactionInProgress is returned when more than one concurrent
+	// transaction is attempted.
+	ErrTransactionInProgress = errors.New("transaction already in progress")
+
+	// ErrTransactionNotStarted is returned when an atomic operation is
+	// requested but the repository hasn't begun a transaction.
+	ErrTransactionNotStarted = errors.New("transaction not started")
+)
+
+type transactor struct {
 	conn   *pgxpool.Pool
+	tx     *pgx.Tx
 	logger *zap.Logger
 }
 
-func NewUserRepo(conn *pgxpool.Pool, logger *zap.Logger) *UserRepo {
-	return &UserRepo{
-		conn:   conn,
-		logger: logger,
+var _ usecases.Transactor = (*transactor)(nil)
+
+func (t *transactor) Begin() error {
+	if t.tx != nil {
+		return ErrTransactionInProgress
+	}
+
+	tx, err := t.conn.Begin(context.Background())
+	if err != nil {
+		t.logger.Error("Failed to Begin Transaction", zap.Error(err))
+		return err
+	}
+	t.tx = &tx
+
+	return nil
+}
+
+func (t *transactor) Commit() error {
+	tx := *t.tx
+	if err := tx.Commit(context.Background()); err != nil {
+		t.logger.Error("Failed to Commit Transaction", zap.Error(err))
+		return err
+	}
+	return nil
+}
+
+func (t *transactor) Rollback() error {
+	tx := *t.tx
+	if err := tx.Rollback(context.Background()); err != nil {
+		t.logger.Error("Failed to Rollback Transaction", zap.Error(err))
+		return err
+	}
+	return nil
+}
+
+type AtomicUserRepo struct {
+	*transactor
+}
+
+var _ usecases.AtomicUserRepository = (*AtomicUserRepo)(nil)
+
+func NewAtomicUserRepoFactory(conn *pgxpool.Pool, logger *zap.Logger) usecases.AtomicUserRepositoryFactory {
+	return func() usecases.AtomicUserRepository {
+		return &AtomicUserRepo{
+			transactor: &transactor{
+				conn:   conn,
+				logger: logger,
+			},
+		}
 	}
 }
 
-func (u *UserRepo) DropUserTable() error {
+func (u *AtomicUserRepo) DropUserTable() error {
+
+	if u.tx == nil {
+		return ErrTransactionNotStarted
+	}
+
 	_, err := u.conn.Exec(context.Background(), "DROP TABLE users;")
 	if err != nil {
 		u.logger.Error("Failed to create table in Database", zap.Error(err))
 		fmt.Println(err)
 	}
+
 	return nil
 }
-func (u *UserRepo) CreateUserTable() error {
+func (u *AtomicUserRepo) CreateUserTable() error {
+	if u.tx == nil {
+		return ErrTransactionNotStarted
+	}
+
 	_, err := u.conn.Exec(context.Background(), "CREATE TABLE users (userid varchar(255), name VARCHAR ( 50 )  NOT NULL,password VARCHAR ( 50 ) NOT NULL, PRIMARY KEY(userid));")
 	if err != nil {
 		u.logger.Error("Failed to create table in Database", zap.Error(err))
 		fmt.Println(err)
 	}
+
 	return nil
 }
 
-func (u *UserRepo) GetIndividual(idd uuid.UUID, wg *sync.WaitGroup, userChan chan<- domain.User) {
+func (u *AtomicUserRepo) GetIndividual(idd uuid.UUID, wg *sync.WaitGroup, userChan chan<- domain.User) {
+
 	defer wg.Done()
 	var user domain.User
 	err := u.conn.QueryRow(context.Background(), "SELECT userid, name, password FROM users WHERE userid = $1", idd).Scan(&user.ID, &user.Name, &user.Password)
@@ -56,7 +126,11 @@ func (u *UserRepo) GetIndividual(idd uuid.UUID, wg *sync.WaitGroup, userChan cha
 	userChan <- user
 }
 
-func (u *UserRepo) GetAll() ([]domain.User, error) {
+func (u *AtomicUserRepo) GetAll() ([]domain.User, error) {
+	if u.tx == nil {
+		return nil, ErrTransactionNotStarted
+	}
+
 	rows, err := u.conn.Query(context.Background(), "SELECT userid FROM users")
 	if err != nil {
 		return nil, err
@@ -91,7 +165,11 @@ func (u *UserRepo) GetAll() ([]domain.User, error) {
 	return users, nil
 }
 
-func (u *UserRepo) GetByID(id uuid.UUID) (*domain.User, error) {
+func (u *AtomicUserRepo) GetByID(id uuid.UUID) (*domain.User, error) {
+	if u.tx == nil {
+		return nil, ErrTransactionNotStarted
+	}
+
 	var user domain.User
 	err := u.conn.QueryRow(context.Background(), "SELECT userid, name, password FROM users WHERE userid = $1", id).Scan(&user.ID, &user.Name, &user.Password)
 
@@ -104,7 +182,10 @@ func (u *UserRepo) GetByID(id uuid.UUID) (*domain.User, error) {
 	return &user, nil
 }
 
-func (u *UserRepo) CreateUser(user *domain.User) error {
+func (u *AtomicUserRepo) CreateUser(user *domain.User) error {
+	if u.tx == nil {
+		return ErrTransactionNotStarted
+	}
 
 	// Added a Check on Password before Creating the User
 	lenght := len(user.Password)
@@ -120,7 +201,11 @@ func (u *UserRepo) CreateUser(user *domain.User) error {
 	return nil
 }
 
-func (u *UserRepo) Withdraw(account *domain.Account, amount int) error {
+func (u *AtomicUserRepo) Withdraw(account *domain.Account, amount int) error {
+	if u.tx == nil {
+		return ErrTransactionNotStarted
+	}
+
 	qry := "UPDATE accounts SET accounts.balance = (accounts.balance - $1) WHERE accounts.userid = $2"
 	_, err := u.conn.Exec(context.Background(), qry, amount, account.UserID)
 	if err != nil {
@@ -130,7 +215,11 @@ func (u *UserRepo) Withdraw(account *domain.Account, amount int) error {
 	return nil
 }
 
-func (u *UserRepo) Deposit(account *domain.Account, amount int) error {
+func (u *AtomicUserRepo) Deposit(account *domain.Account, amount int) error {
+	if u.tx == nil {
+		return ErrTransactionNotStarted
+	}
+
 	/*
 		UPDATE product
 		SET net_price = price - price * discount
